09-Security-Projects/authentication-system/requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Limiter==3.5.0

# Security & Authentication
bcrypt==4.1.2
PyJWT==2.8.0
pyotp==2.9.0
cryptography==41.0.7
python-dotenv==1.0.0

# Database
Flask-Migrate==4.0.5
redis==5.0.1

# Validation
email-validator==2.1.0
pydantic==2.5.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development
python-dotenv==1.0.0
gunicorn==21.2.0
